+++ apstools/devices/tests/test_aps_cycle.py
@@ -16,6 +16,7 @@ def using_APS_workstation():
     return hostname.lower().endswith(".aps.anl.gov")
 
 
+@pytest.mark.skip("TODO issue 1122")
 def test_ApsCycleDM():
     signal = ApsCycleDM(name="signal")
     assert signal.connected
@@ -32,6 +33,7 @@ def test_ApsCycleDM():
         assert int(cycle.split("-")[0]) > 2020
 
 
+@pytest.mark.skip("TODO issue 1122")
 def test_ApsCycleDB():
     assert aps_cycle.YAML_CYCLE_FILE.exists()